### 2025-08-23 #258 @Ematrion
RSTT Submission
Submitting Author: David Bucher (@Ematrion)
Package RSTT
One-Line Description of Package: Competition simulation tool to evaluate ranking
Repository Link:  https://github.com/Ematrion/rstt
Version submitted: v0.6.7-alpha
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

-
Package enable sport/games simulation providing different tournament format, state of the art rating system and probabilistic model. It is build around the interaction of a ranking, used to seed competition, and the resulting games, used to update the ranking. It allows users to compared trained ranking with the simulation model, and answer question such has 'Does the ranking have a unique stationary distribution'.

It aims at enabling simulation based research in the context of competition.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

I believe it fits the "Scope for packages that support analytics, statistics and modeling":
It helps producing well defined, yet complex synthetic dataset with a precise and short syntax, 20 lines of code or less for millions of games. It is possible to reproduce others research (check tutorial_3) in the field.

           - data scientist testing rating system for large video game title or sport leagues.
           - game developer who wants to make simple test.
           - teaching tool to illustrate ranking behaviour in an intuitive context.
           - mathematician studying Markov process and stationery distribution.
           - tournament organiser with care for matching and seeding methods

I know there is a paper on a tool call MMBench: https://dl.acm.org/doi/10.1145/3539597.3573023 . I could not access it. It seems limited to its own component while RSTT allows user to define and integrate custom model. It does not seems to address the interplay between ranking and game generation project. Does not seems to tune the initial ranking state.

  - If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or `@tag` the editor you contacted: https://github.com/pyOpenSci/software-submission/issues/255

       @eliotwrobson

https://doi.org/10.5281/zenodo.16926606
