### 2025-08-27 833d0cb
gpr tutotrial
BUG FIX:
+ eventscoring.rate method when player did not participate in events
+ successranking.update method to support relevance args
with tests suits
