### 2025-08-31 24cb718
start of GPR tutorial
- json data
- usage of StrEnum WIP
- Model implementation
- basic notebooks runs
