### 2025-08-31 eaa6da0
Update test_eventscoring.py
