### 2025-08-31 2aac6ac
Update issue template
ADD: Feature Request
