### 2025-09-15 18e1adf
update tuto GPR
