+++ tutorials/4_Modeling_GPR/exercises/gpr/__init__.py
@@ -1,0 +1,4 @@
+from .ratingsystem import RegionalRatings
+from .inference import GPRElo
+from .observer import GlobalHandler
+from .globalranking import GlobalRanking
