+++ tutorials/4_Modeling_GPR/exercises/gpr/ratingsystem.py
@@ -1,0 +1,9 @@
+from typing import Any
+
+
+class RegionalRatings():
+    def __init__(self):
+        ...
+
+    def ordinal(self, rating: Any) -> float:
+        ...
