+++ tutorials/4_Modeling_GPR/project/gpr/__init__.py
@@ -1,0 +1,5 @@
+from .ratingsystem import RegionalRatings
+from .observer import GlobalHandler, LeagueGBG
+from .utils import GameHistory, LeagueSystem
+
+from .globalranking import GlobalRanking
