### 2025-09-18 c3b9ef5
ADD:     + settings.py file with model specifications
FIX:
    + lolteam typing with SPlayer and BasicPlayer

DELETE:
    - experiments.one_round_robin
