+++ tutorials/4_Modeling_GPR/exercises/model/meta.py
@@ -1,1 +1,8 @@
 
+
+class MetaData():
+    def __init__(self) -> None:
+        ...
+
+    def update(self) -> None:
+        ...
