### 2025-11-12 fea85c6
fix in game data extraction
(global package) changed update keyword:
ratings_groups -> rating_group

deadly typo that made the package incompatible with other services
added y test_elo in rankings to ensure the update workflow
