+++ rstt/utils/observer.py
@@ -12,7 +12,7 @@ PLAYER = 'player'
 PLAYERS = 'players'
 TEAM = 'team'
 TEAMS = 'teams'
-RATINGS_GROUPS = 'ratings_groups'
+RATINGS_GROUPS = 'rating_groups'
 RATINGS_OPPONENTS = 'ratings_opponents'
 RATING = 'rating'
 RATING1 = 'rating1'
