+++ rstt/ranking/inferer/elo.py
@@ -23,12 +23,12 @@ class Elo:
         # QUEST: should the base implementation support distribution function as parameters
 
     @typechecked
-    def rate(self, ratings_groups: list[list[float]], scores: list[float], *args, **kwars) -> list[list[float]]:
+    def rate(self, rating_groups: list[list[float]], scores: list[float], *args, **kwars) -> list[list[float]]:
         """Rate method for elo
 
         Parameters
         ----------
-        ratings_groups : List[List[float]]
+        rating_groups : List[List[float]]
             Elo ratings formated by teams, for example [[elo_player1], [elo_player2]].
         scores : List[float]
             corresponding scores of the ratings, for example [[1.0],[0.0]] assuming player1 won the duel.
@@ -43,10 +43,10 @@ class Elo:
         !!! NOBUG:
             Take great care when reading the code, both calls are valid:
             - updating ratings based on one game score:
-                elo.rate(ratings_groups=[[1500],[1600]], scores=[0.0, 1.0])
+                elo.rate(rating_groups=[[1500],[1600]], scores=[0.0, 1.0])
                 -> output two ratings
             - updating one player rating based on one game score:
-                elo.rate(ratings_groups=[[1500], [1600]], scores=[0.0])
+                elo.rate(rating_groups=[[1500], [1600]], scores=[0.0])
                 -> output one rating
                 
             This can be confusing as hell, however the requierements are:
@@ -56,22 +56,22 @@ class Elo:
         '''
 
         # Deal with bad function calls
-        if len(ratings_groups) != 2:
-            msg = f"Expect two ratings groups, got {len(ratings_groups)}"
+        if len(rating_groups) != 2:
+            msg = f"Expect two ratings groups, got {len(rating_groups)}"
             raise ValueError(msg)
-        if len(ratings_groups[0]) != 1:
-            msg = f"Expect only one rating in the first ratings group, got {len(ratings_groups[0])}"
+        if len(rating_groups[0]) != 1:
+            msg = f"Expect only one rating in the first ratings group, got {len(rating_groups[0])}"
             raise ValueError(msg)
-        if len(ratings_groups[1]) == 1 and len(scores) not in [1, 2]:
+        if len(rating_groups[1]) == 1 and len(scores) not in [1, 2]:
             msg = f"For 1-versus-1 update, Elo Expect \'scores\' of len 2, received {len(scores)}"
             raise ValueError(msg)
-        if len(ratings_groups[1]) != 1 and len(ratings_groups[1]) != len(scores):
-            msg = f"Incompatible args call, 2nd ratings group must be of length equal to the scores, received {len(ratings_groups[1])} and {len(scores)}"
+        if len(rating_groups[1]) != 1 and len(rating_groups[1]) != len(scores):
+            msg = f"Incompatible args call, 2nd ratings group must be of length equal to the scores, received {len(rating_groups[1])} and {len(scores)}"
             raise ValueError(msg)
 
-        if len(ratings_groups[0]) == len(ratings_groups[1]) == 1 and len(scores) == 2:
+        if len(rating_groups[0]) == len(rating_groups[1]) == 1 and len(scores) == 2:
             # one 1-versus-1 case
-            [[r1], [r2]] = ratings_groups
+            [[r1], [r2]] = rating_groups
             [s1, s2] = scores
             new_rating1 = self.post_rating(
                 prior_rating=r1, ratings_opponents=[r2], scores=[s1])
@@ -80,7 +80,7 @@ class Elo:
             return [[new_rating1], [new_rating2]]
         else:
             # many 1-versus-1 case
-            [[r1], rs] = ratings_groups
+            [[r1], rs] = rating_groups
[…]
