+++ tests/ranking/test_glicko2.py
@@ -58,7 +58,7 @@ def glicko2():
     return BasicGlicko2(name='test-glicko-2', tau=0.5)
 
 
-def test_glicko2__with_paper_example(glicko2, player, p1, p2, p3, games):
+def test_glicko2_with_paper_example(glicko2, player, p1, p2, p3, games):
     # the test: glicko2 user interface
     for p in [player, p1, p2, p3]:
         glicko2.set_rating(p.player, p.prior)
