+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "rstt"
-version = "0.6.7"
+version = "0.6.7.1"
 description = "Ranking Simulation Testing Tool"
 authors = [
     {name = "Bucher David", email = "davidbucherdev@gmail.com"}
