### 2025-12-05 #258 @Ematrion
Hi @eliotwrobson, just wanted to have some news about progress, how is the going, how can I help?

Also, what is the policy about updates and code changes?
There is an almost ready tutorial I would like to upload and it comes with a few bug fixes.
What about feature additions ?
Basically to what extend can I edit the code base?
