### 2026-01-20 #258 @amyheather
Hi @Ematrion - I'm not sure if there's an official approach - but I likewise have a package at same stage, and am pretty sure its fine to continue updating your codebase while at seeking-editor stage. My approach has been simply to edit the "version submitted" in my initial GitHub issue, and to also post a comment mentioning that I have done so e.g. "Edit: Updated version submitted (v0.6.0 --> v0.7.0)." (can see here - https://github.com/pyOpenSci/software-submission/issues/257 ) - hope that helps! 😁
