### 2026-02-01 #258 @crhea93
@Ematrion Hi and thank you for your submission! I'll be the editor in chief going forward on this review. The first step is for me to complete a pre-review checklist in which I will ensure that the codebase meets the current standards for PyOpenSci. Afterwards, you will be able to make any corrections that I request :) Afterwards, I'll find an editor :D

You can absolutely add features to the code base while under review! If there are any major changes, however, please mention them in this issue and understand that the review may take a little longer since the reviewers will have to go over those as well :)
