### 2026-02-04 d283b65
COF
+ CODE_OF_CONDUCT.md, Discord for enforcement
+ Discosure section in the readme for AI usage
+ installation and Examle link to documentation in the readme
