### 2026-02-04 #258 @Ematrion
Hi @crhea93 , thank you very much for feedback.
I included a COF, an AI disclosure at the end of the readme and added some link to the docs installation instruction and example explanation.

As for the Onboard survey, I remember having fill out a google form at submission (right before or after).
